+++ .github/scripts/UnityEngine_stubs.cs
@@ -7,6 +7,7 @@ namespace UnityEngine
     public abstract class MonoBehaviour : Behaviour
     {
         public Coroutine StartCoroutine(IEnumerator routine) => null;
+        public void StopCoroutine(Coroutine routine) { }
         public void StopAllCoroutines() { }
     }
     public class Behaviour : Component
@@ -18,6 +19,7 @@ namespace UnityEngine
         public T GetComponent<T>() where T : class => null;
         public Component GetComponent(Type type) => null;
         public T GetComponentInChildren<T>() where T : class => null;
+        public T GetComponentInParent<T>() where T : class => null;
         public GameObject gameObject { get; }
         public Transform transform { get; }
     }
@@ -25,11 +27,23 @@ namespace UnityEngine
     {
         public GameObject() { }
         public GameObject(string name) { }
+        public string name { get; set; } = "";
         public T AddComponent<T>() where T : Component => null;
+        public T GetComponent<T>() where T : class => null;
         public Component AddComponent(Type componentType) => null;
         public T[] GetComponentsInChildren<T>() where T : class => null;
+        public T[] GetComponentsInChildren<T>(bool includeInactive) where T : class => null;
+        public T GetComponentInChildren<T>() where T : class => null;
         public Transform transform { get; }
         public bool activeInHierarchy { get; }
+        public void SetActive(bool value) { }
+        public static GameObject CreatePrimitive(PrimitiveType type) => new GameObject();
+    }
+
+    public enum PrimitiveType
+    {
+        Sphere = 0,
+        Capsule = 2,
     }
     public class Transform : Component
     {
@@ -37,10 +51,19 @@ namespace UnityEngine
         public Vector3 forward { get; }
         public Vector3 right { get; }
         public Vector3 localPosition { get; set; }
+        public Vector3 localScale { get; set; }
+        public Vector3 lossyScale { get; set; }
         public Vector3 localEulerAngles { get; set; }
+        public Quaternion rotation { get; set; }
+        public Quaternion localRotation { get; set; }
         public Transform parent { get; set; }
+        public Transform root { get; }
         public void SetParent(Transform parent) { }
         public void SetParent(Transform parent, bool worldPositionStays) { }
+        public int childCount { get; }
+        public Transform GetChild(int index) => null;
+        public Vector3 InverseTransformPoint(Vector3 position) => position;
+        public Vector3 TransformPoint(Vector3 position) => position;
         public bool IsChildOf(Transform potentialParent)
         {
             var t = this;
@@ -54,11 +77,21 @@ namespace UnityEngine
     }
     public class Object
     {
+        public string name { get; set; } = "";
         public static T FindObjectOfType<T>() where T : Object => null;
         public static T[] FindObjectsOfType<T>() where T : Object => null;
         public static void Destroy(Object obj, float t = 0f) { }
         public static void DontDestroyOnLoad(Object obj) { }
+        public static Object Instantiate(Object original) => original;
+        public static Object Instantiate(Object original, Vector3 position, Quaternion rotation) => original;
+        public static Object Instantiate(Object original, Transform parent) => original;
         public static implicit operator bool(Object exists) { return exists != null; }
+    }
+    public struct Quaternion
+    {
+        public static Quaternion LookRotation(Vector3 forward) => new Quaternion();
+        public static Quaternion Inverse(Quaternion rotation) => rotation;
+        public static Quaternion operator *(Quaternion lhs, Quaternion rhs) => lhs;
     }
     public struct Vector2
     {
@@ -81,10 +114,12 @@ namespace UnityEngine
         public float x, y, z;
         public Vector3 normalized => this;
         public float magnitude => 0f;
+        public float sqrMagnitude => 0f;
         public static Vector3 forward => new Vector3();
         public static Vector3 right => new Vector3();
         public static Vector3 zero => new Vector3();
         public static float Distance(Vector3 a, Vector3 b) => 0f;
+        public static Vector3 Lerp(Vector3 a, Vector3 b, float t) => a;
         public Vector3(float x, float y, float z) { this.x = x; this.y = y; this.z = z; }
         public static Vector3 operator *(Vector3 v, float s) => v;
         public static Vector3 operator -(Vector3 v) => v;
@@ -96,8 +131,42 @@ namespace UnityEngine
         public static Camera main { get; }
         public new Transform transform { get; }
     }
+    public class Material : Object { }
+
+    public struct Bounds
+    {
+        public Vector3 min;
+        public Vector3 max;
+        public Bounds(Vector3 center, Vector3 size)
+        {
+            min = center - size * 0.5f;
+            max = center + size * 0.5f;
+        }
+    }
+
+    public class Mesh : Object
+    {
+        public int vertexCount { get; set; }
+        public Bounds bounds { get; set; }
+    }
+
+    public class Renderer : Behaviour
+    {
+        public Material[] sharedMaterials { get; set; }
+    }
+
+    public class MeshFilter : Component
+    {
+        public Mesh sharedMesh { get; set; }
+        public Mesh mesh { get; set; }
+    }
+
+    public class MeshRenderer : Renderer { }
+
     public class AudioSource : Behaviour
     {
+        public static void PlayClipAtPoint(AudioClip clip, Vector3 position, float volume) { }
+
         public AudioClip clip { get; set; }
         public float volume { get; set; }
         public float pitch { get; set; }
@@ -158,6 +227,12 @@ namespace UnityEngine
         public static float Sin(float f) => 0f;
         public static float Sqrt(float f) => 0f;
         public static float Clamp01(float value) => value;
+        public static float Pow(float x, float y) => (float)Math.Pow(x, y);
+    }
+
+    public class Animator : Behaviour
+    {
+        public void SetTrigger(string name) { }
     }
 
     public static class Physics
@@ -182,7 +257,10 @@ namespace UnityEngine
         public Vector3 point { get; }
     }
 
-    public class Collider : Component { }
+    public class Collider : Behaviour
+    {
+        public bool isTrigger { get; set; }
+    }
 
     public enum QueryTriggerInteraction
     {